% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_purchase_patterns.R
\name{compute_purchase_frequency}
\alias{compute_purchase_frequency}
\title{Function to compute purchase frequency}
\usage{
compute_purchase_frequency(purchases, gap = 30, use_pills_per_pack_only = TRUE)
}
\arguments{
\item{purchases}{data.frame of purchases for a single VNR as returned by get_drug_purchases. 
if use_pills_per_pack_only Uses max(PackageSize, DDDPerPack) + gap else use packageSize + gap to determine which adjacent purchases are considered part of the same treatment interval}

\item{gap}{maximum permissible gap between purchases}

\item{use_pills_per_pack_only}{logical, whether to use only PackageSize + gap to determine treatment intervals default TRUE}
}
\value{
data.frame of purchase intervals
}
\description{
For a given set of purchases for a single VNR, compute the intervals between purchases for each individual. Purchases are considered part of the same treatment interval if they are within max(PackageSize, DDDPerPack) + gap days of each other.
}
