% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{.apply_summary}
\alias{.apply_summary}
\title{Internal helper function to apply summary function to lab measurements for a given period and check results}
\usage{
.apply_summary(lab_measurements, summary_function)
}
\arguments{
\item{lab_measurements}{data frame of lab measurements for an individual}

\item{summary_function}{function to summarize lab measurements (e.g., summary_median, summary_closest_to_drug)}
}
\description{
Internal helper function to apply summary function to lab measurements for a given period and check results
}
