% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{filter_outliers_mad}
\alias{filter_outliers_mad}
\title{Filter Outliers using Median Absolute Deviation (MAD)}
\usage{
filter_outliers_mad(x, th = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{th}{The threshold for MAD. A higher value is more conservative. Defaults to 5.}
}
\value{
A numeric vector with outliers removed.
}
\description{
Removes outliers from a numeric vector based on the Median Absolute Deviation (MAD).
This method is generally more robust than standard deviation-based outlier removal.
}
