% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc_mapping.R
\name{get_atc_relationships}
\alias{get_atc_relationships}
\title{Get All Related ATC Codes}
\usage{
get_atc_relationships(atc_code, mappings = NULL)
}
\arguments{
\item{atc_code}{Single ATC code to query}

\item{mappings}{Optional pre-loaded mappings (if NULL, will load from cache)}
}
\value{
List with detailed mapping information
}
\description{
Returns all codes related to a given ATC code (historical, current, and related)
}

