% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_first_purchase}
\alias{get_first_purchase}
\title{Get first drug purchase from FinnGen data}
\usage{
get_first_purchase(
  conn,
  druglist,
  finngen_ids = NULL,
  use_only_reimbursement = FALSE,
  use_atc_mapping = TRUE,
  lazy = FALSE
)
}
\arguments{
\item{conn}{finngen data connection object}

\item{druglist}{vector of drug ATC codes. The ATC codes are matched with the first part of the code (e.g. A01*)}

\item{finngen_ids}{vector of FINNGENIDs to filter the data. leave empty to get all}

\item{use_only_reimbursement}{logical, if TRUE, use only reimbursement data (default FALSE) and combine reimbursement and delivery data if available in conn}

\item{use_atc_mapping}{logical, if TRUE, expand ATC codes to include historical alterations (default TRUE)}

\item{lazy}{logical, if TRUE, return a lazy tbl object}
}
\value{
data frame with first drug purchases for each FINNGENID
}
\description{
Get first drug purchase from FinnGen data
}
