% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_lab_measurements}
\alias{get_lab_measurements}
\title{Get lab measurements from FinnGen data}
\usage{
get_lab_measurements(
  all_labs,
  lablist,
  require_values = TRUE,
  return_cols = c("FINNGENID", "OMOP_CONCEPT_ID", "EVENT_AGE",
    "MEASUREMENT_VALUE_HARMONIZED"),
  finngen_ids = NULL,
  lazy = FALSE,
  covariates = NULL,
  covariate_cols = NULL
)
}
\arguments{
\item{all_labs}{data frame with lab measurements}

\item{lablist}{vector of lab measurement concept IDs}

\item{require_values}{logical, if TRUE, only return rows with non-missing MEASUREMENT_VALUE_HARMONIZED}

\item{return_cols}{vector of column names to return}

\item{finngen_ids}{vector of FINNGENIDs to filter the data}

\item{lazy}{logical, if TRUE, return a lazy tbl object}

\item{covariates}{Optional data frame with covariate data (e.g., sex, age at death)}

\item{covariate_cols}{Optional character vector of column names to add from the covariates data frame}
}
\value{
data frame with lab measurements
}
\description{
Get lab measurements from FinnGen data
}
