% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{inverse_rank_normalize}
\alias{inverse_rank_normalize}
\title{Inverse Rank Normalize Values}
\usage{
inverse_rank_normalize(x)
}
\arguments{
\item{x}{Numeric vector to normalize}
}
\value{
Inverse rank normalized vector with same length as input
}
\description{
Performs inverse rank normalization on a numeric vector
}
