% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc_mapping.R
\name{load_atc_mappings}
\alias{load_atc_mappings}
\title{Load ATC Mappings}
\usage{
load_atc_mappings(custom_file = NULL)
}
\arguments{
\item{custom_file}{Optional path to a custom mapping file (for testing or updates)}
}
\value{
List of ATC mappings
}
\description{
Loads ATC mappings from packaged JSON file
}

