% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_purchase_patterns.R
\name{parallel_compute_purchase_frequencies_for_VNRs}
\alias{parallel_compute_purchase_frequencies_for_VNRs}
\title{Function to compute purchase frequencies for all VNRs in parallel.}
\usage{
parallel_compute_purchase_frequencies_for_VNRs(
  data,
  gap,
  use_pills_per_pack_only = TRUE,
  n_workers = NULL
)
}
\arguments{
\item{data}{data.frame of purchases for multiple VNRs as returned by get_drug_purchases}

\item{gap}{maximum permissible gap between purchases to consider them part of the same treatment interval}

\item{use_pills_per_pack_only}{logical, whether to use only PackageSize + gap to determine treatment intervals default TRUE}

\item{n_workers}{number of parallel workers to use, default NULL uses detectCores() - 1}
}
\value{
data.frame of purchase intervals across all VNRs
}
\description{
Calls compute_purchase_frequency for each VNR, check that function for details of computation
}
