% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Quantile Normalize Values}
\usage{
quantile_normalize(x)
}
\arguments{
\item{x}{Numeric vector to normalize}
}
\value{
Quantile normalized vector with same length as input
}
\description{
Performs quantile normalization on a numeric vector
}
