% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{summarize_drug_purchases_upset}
\alias{summarize_drug_purchases_upset}
\title{Generate an UpSet plot of drug purchase combinations}
\usage{
summarize_drug_purchases_upset(drug_response, out_file_prefix)
}
\arguments{
\item{drug_response}{A `drug.response` object created by `create_drug_response`.}

\item{out_file_prefix}{A string to use as the prefix for the output PDF file.}
}
\description{
Creates and saves an UpSet plot to visualize the overlap of purchased drug ATC codes.
}
