% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{summary_closest_to_drug}
\alias{summary_closest_to_drug}
\title{Summary function to get lab value closest to drug initiation, to be used in create_drug_response call as summary_functions parameter}
\usage{
summary_closest_to_drug(lab_values)
}
\arguments{
\item{lab_values}{data frame of lab measurements for an individual}
}
\description{
Summary function to get lab value closest to drug initiation, to be used in create_drug_response call as summary_functions parameter
}
