% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{summary_min}
\alias{summary_min}
\title{Summary function to get minimum lab value of responses, to be used in create_drug_response call as summary_functions parameter}
\usage{
summary_min(lab_values)
}
\arguments{
\item{lab_values}{data frame of lab measurements for an individual}
}
\description{
Summary function to get minimum lab value of responses, to be used in create_drug_response call as summary_functions parameter
}
