% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_functions.R
\name{create_drug_response}
\alias{create_drug_response}
\title{Create drug response object}
\usage{
create_drug_response(
  conn,
  lablist,
  druglist,
  before_period,
  after_period,
  summary_function = median,
  filter_min_max = c(-Inf, Inf),
  finngen_ids = NULL,
  use_lab_free_text_values = TRUE,
  use_only_reimbursement_drugs = FALSE
)
}
\arguments{
\item{conn}{fg_data_connection object}

\item{lablist}{vector of lab measurement concept IDs}

\item{druglist}{vector of drug ATC codes. The ATC codes are matched with the first part of the code (e.g. A01*).}

\item{before_period}{vector with two elements, start and end of the before period}

\item{after_period}{vector with two elements, start and end of the after period}

\item{summary_function}{function to summarize the lab measurements (default is median, use mean for mean etc.)}

\item{filter_min_max}{vector of length 2, min and max values to filter lab measurements (default c(-Inf, Inf))}

\item{finngen_ids}{vector of FINNGENIDs to filter the data}

\item{use_lab_free_text_values}{logical, if TRUE, use lab measurements combining reported values and those from free text values (default TRUE)}

\item{use_only_reimbursement_drugs}{logical, if TRUE, use only reimbursement data (i.e. longitudinal data from purchase registry) (default FALSE)}
}
\value{
drug.response object
}
\description{
Create drug response object
}
