% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{create_drug_response}
\alias{create_drug_response}
\title{Create drug response object}
\usage{
create_drug_response(
  conn,
  lablist,
  druglist,
  before_period,
  after_period,
<<<<<<< HEAD
  summary_function = median,
  filter_min_max = c(-Inf, Inf),
  finngen_ids = NULL,
  use_lab_free_text_values = TRUE,
  use_only_reimbursement_drugs = FALSE
=======
  finngen_ids = NULL,
  remove_outliers_sd = NULL
>>>>>>> master
)
}
\arguments{
\item{conn}{A `fg_data_connection` object containing the data sources}

\item{lablist}{A character vector of OMOP concept IDs for the labs of interest}

\item{druglist}{A character vector of ATC drug codes}

\item{before_period}{A numeric vector of length 2 defining the before period (e.g., c(-1, 0) for 1 year to 0 before drug)}

\item{after_period}{A numeric vector of length 2 defining the after period (e.g., c(0.1, 1) for 0.1 to 1 year after drug)}

<<<<<<< HEAD
\item{summary_function}{function to summarize the lab measurements (default is median, use mean for mean etc.)}

\item{filter_min_max}{vector of length 2, min and max values to filter lab measurements (default c(-Inf, Inf))}

\item{finngen_ids}{vector of FINNGENIDs to filter the data}

\item{use_lab_free_text_values}{logical, if TRUE, use lab measurements combining reported values and those from free text values (default TRUE)}

\item{use_only_reimbursement_drugs}{logical, if TRUE, use only reimbursement data (i.e. longitudinal data from purchase registry) (default FALSE)}
=======
\item{finngen_ids}{Optional character vector of FINNGENIDs to restrict analysis}

\item{remove_outliers_sd}{Optional numeric value (1-6) to remove outliers based on standard deviations}
>>>>>>> master
}
\value{
drug.response object
}
\description{
Create drug response object
}
