% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_functions.R
\name{create_drug_response}
\alias{create_drug_response}
\title{Create drug response object}
\usage{
create_drug_response(
  conn,
  lablist,
  druglist,
  before_period,
  after_period,
  finngen_ids = NULL,
  remove_outliers_sd = NULL
)
}
\arguments{
\item{conn}{fg_data_connection object}

\item{lablist}{vector of lab measurement concept IDs}

\item{druglist}{vector of drug ATC codes. The ATC codes are matched with the first part of the code (e.g. A01*).}

\item{before_period}{vector with two elements, start and end of the before period}

\item{after_period}{vector with two elements, start and end of the after period}

\item{finngen_ids}{vector of FINNGENIDs to filter the data}

\item{remove_outliers_sd}{integer, if defined, remove outliers from the lab measurements. The value is the number of standard deviations from the mean to define an outlier (i.e. 1, 2, 3, 4, 5, 6)}
}
\value{
drug.response object
}
\description{
Create drug response object
}
