% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{create_drug_response}
\alias{create_drug_response}
\title{Create drug response object}
\usage{
create_drug_response(
  conn = NULL,
  lablist = NULL,
  druglist = NULL,
  before_period = NULL,
  after_period = NULL,
  finngen_ids = NULL,
  remove_outliers_sd = NULL,
)
}
\arguments{
\item{conn}{Connection object or kanta data frame (for legacy support)}

\item{lablist}{Lab list or phenos data frame (for legacy support)}

\item{druglist}{Drug list}

\item{before_period}{Before period}

\item{after_period}{After period}

\item{finngen_ids}{Finngen IDs}

\item{remove_outliers_sd}{Remove outliers SD}

}
\value{
drug.response object
}
\description{
Create drug response object
}
