% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_core.R
\name{generate_response_summary}
\alias{generate_response_summary}
\title{Generate response summary}
\usage{
generate_response_summary(
  lab_measurements,
  drug_purchases = NULL,
  before_period,
  after_period,
  summary_functions = list(summary_median, summary_median)
)
}
\arguments{
\item{lab_measurements}{data frame with lab measurements and indicator of time to first drug}

\item{drug_purchases}{data frame with drug purchases and indicator of time to first drug}

\item{before_period}{vector with two elements, start and end of the before period}

\item{after_period}{vector with two elements, start and end of the after period}

\item{summary_functions}{A list of two functions to summarize the lab measurements (default is list(summary_median, summary_median)). The first function is used for baseline summary and the second for followup summary.}
}
\value{
data frame with response summary
}
\description{
Generate response summary
}
