% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_drug_purchases}
\alias{get_drug_purchases}
\title{Get drug purchases from FinnGen data}
\usage{
get_drug_purchases(
  all_phenos,
  druglist,
  finngen_ids = NULL,
  return_cols = c("FINNGENID", "EVENT_AGE", ATC = "CODE1", REIMB_CODE = "CODE2", VNR =
    "CODE3", N_PACKS = "CODE4"),
  lazy = FALSE
)
}
\arguments{
\item{all_phenos}{data frame with drug purchases}

\item{druglist}{vector of drug ATC codes. The ATC codes are matched with the first part of the code (e.g. A01*)}

\item{finngen_ids}{vector of FINNGENIDs to filter the data. leave empty to get all}

\item{return_cols}{vector of column names to return}

\item{lazy}{logical, if TRUE, return a lazy tbl object}
}
\value{
data frame with drug purchases
}
\description{
Get drug purchases from FinnGen data
}
