% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_response_functions.R
\name{get_drug_purchases}
\alias{get_drug_purchases}
\title{Get drug purchases from FinnGen data}
\usage{
get_drug_purchases(
  conn,
  druglist,
  finngen_ids = NULL,
  use_only_reimbursement = FALSE,
  lazy = FALSE
)
}
\arguments{
\item{conn}{finngen data connection object}

\item{druglist}{vector of drug ATC codes. The ATC codes are matched with the first part of the code (e.g. A01*)}

\item{finngen_ids}{vector of FINNGENIDs to filter the data. leave empty to get all}

\item{use_only_reimbursement}{logical, if TRUE, use only reimbursement data (default FALSE) and combine reimbursement and delivery data if available in conn}

\item{lazy}{logical, if TRUE, return a lazy tbl object}
}
\value{
data frame with drug purchases
}
\description{
Get drug purchases from FinnGen data
}
