% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{join_covariates}
\alias{join_covariates}
\title{Join Covariates to Any Data Frame}
\usage{
join_covariates(data, covariates, covariate_cols)
}
\arguments{
\item{data}{Data frame with FINNGENID column}

\item{covariates}{Data frame or lazy table containing covariate data}

\item{covariate_cols}{Character vector of column names to join from the covariates table}
}
\value{
Data frame with joined covariates
}
\description{
Generic helper function to join covariate data to any data frame with FINNGENID.
}
\examples{
# Join covariates to drug response data
response_with_covariates <- join_covariates(
  data = drug_response$responses,
  covariates = conn$cov_pheno,
  covariate_cols = c("SEX", "AGE_AT_DEATH_OR_END_OF_FOLLOWUP")
)
}
