% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_individual_trajectory}
\alias{visualize_individual_trajectory}
\title{Visualize Individual Trajectory of Lab Values}
\usage{
visualize_individual_trajectory(drug_response, finngen_id)
}
\arguments{
\item{drug_response}{A `drug.response` object containing the data to visualize.}

\item{finngen_id}{The FINNGEN ID of the individual to visualize.}
}
\value{
A `ggplot` object showing the trajectory of lab values for the specified individual.
}
\description{
Plots the trajectory of lab values for a specific individual (FINNGEN
ID) over time, with time to drug purchase on the x-axis and lab measurement on the y-axis.
The plot includes a dashed vertical line at the time of drug purchase (time_to_drug = 0) and dotted vertical lines for each drug purchase during follow-up.
}
